bovitech-chatbot-main/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
Django==5.2.12
django-cors-headers==4.9.0
ffmpeg-python==0.2.0
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
future==1.0.0
gast==0.7.0
google-pasta==0.2.0
groq==1.1.1
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
libclang==18.1.1
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
openai-whisper==20250625
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
portalocker==3.2.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.1
regex==2026.2.28
requests==2.33.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wrapt==2.1.2
imageio-ffmpeg>=0.5.1
edge-tts>=7.0.0
